[fix] comments and linting
