+++ lib/preboots/scheduler.js
@@ -1,8 +1,7 @@
-/* eslint max-params: 0*/
 const EE = require('events');
 const async = require('async');
 const once = require('one-time');
-const semver = require('semver');
+const Lookup = require('../lookup');
 const nsqStream = require('nsq-stream');
 const from = require('from2');
 const util = require('util');
@@ -13,8 +12,15 @@ util.inherits(Scheduler, EE);
 
 /**
  * Scheduler
+ * @public
  * @constructor
  * @param {Object} options - Configuration object
+ * @param {Object} options.log - Logger
+ * @param {Object} options.nsq - NSQ object
+ * @param {String} options.topic - NSQ topic to dispatch onto
+ * @param {Object} options.models - Data models
+ * @param {Number} [options.concurrency=5] - Number of jobs to run concurrently
+ * @param {Number} [options.interval=3600000] - Default interval between catch-up jobs in ms
  */
 function Scheduler(options) {
   EE.call(this);
@@ -30,14 +36,14 @@ function Scheduler(options) {
 
 /**
  * Set an interval to schedule catch up jobs for the given env
+ * @public
  * @function setInterval
  * @param {String} env - Environment
- * @param {Number} time - milliseconds of interval
- * @returns {undefined}
+ * @param {Number} [time] - milliseconds of interval
  */
 Scheduler.prototype.setInterval = function interval(env, time) {
   time = time || this.defaultInterval;
-  const int = setInterval(() => {
+  const interv = setInterval(() => {
     this.emit('schedule');
     this.schedule(env, (err, counts) => {
       if (err) return this.log.error('Failed to schedule jobs for %s', env, {
@@ -49,13 +55,14 @@ Scheduler.prototype.setInterval = function interval(env, time) {
     });
   }, time);
 
-  this.intervals.set(env, int);
+  this.intervals.set(env, interv);
 };
 
 /**
  * Clear the given setIntervals for the given env or all of them
+ * @public
  * @function clear
- * @param {String} env - Optional env to pass
+ * @param {String} [env] - Optional env to pass
  * @returns {undefined}
  */
 Scheduler.prototype.clear = function clear(env) {
@@ -67,23 +74,21 @@ Scheduler.prototype.clear = function clear(env) {
 
 /**
  * Core clearing of setIntervals that cleans up map value as well
+ * @private
  * @function _clear
- * @api private
  * @param {String} env - Environment
- * @returns {undefined}
  */
 Scheduler.prototype._clear = function (env) {
-  clearInterval(this.intervals.get(env));
+  if (this.intervals.has(env)) clearInterval(this.intervals.get(env));
   this.intervals.delete(env);
 };
 
 /**
  * Schedule catch up jobs for the given environment over nsq
+ * @public
  * @function schedule
- * @api public
  * @param {String} env - Environment to schedule for
  * @param {Function} callback - Continuation to call when completed
- * @returns {undefined}
  */
 Scheduler.prototype.schedule = function schedule(env, callback) {
   const counts = {};
@@ -102,11 +107,11 @@ Scheduler.prototype.schedule = function schedule(env, callback) {
   this.packages((err, packages) => {
     if (err) return callback(err);
     async.eachLimit(packages, this.conc, (pkg, next) => {
-      const fn = once(next);
+      const cb = once(next);
       const name = pkg.name;
       this.models.BuildHead.findAll({ name, env }, (err, heads) => {
-        if (err) return fn(err);
-        if (!heads || !heads.length) return fn();
+        if (err) return cb(err);
+        if (!heads || !heads.length) return cb();
 
         const lookup = new Lookup(pkg, heads);
         const writer = nsqStream.createWriteStream(this.nsq.writer, this.topic);
@@ -118,8 +123,8 @@ Scheduler.prototype.schedule = function schedule(env, callback) {
 
         from.obj(missing)
           .pipe(writer)
-          .on('error', fn)
-          .once('finish', fn);
+          .on('error', cb)
+          .once('finish', cb);
       });
     }, done);
   });
@@ -128,95 +133,20 @@ Scheduler.prototype.schedule = function schedule(env, callback) {
 /**
  * Fetch all packages we currently have stored using our special cache table
  * for fast fetches
+ * @public
  * @function packages
- * @api public
- * @param {Function} fn - Continuation function to call when completed
- * @returns {undefined}
+ * @param {Function} callback - Continuation function to call when completed
  */
-Scheduler.prototype.packages = function packs(fn) {
-  this.models.PackageCache.findAll({ partitioner: 'cached' }, fn);
+Scheduler.prototype.packages = function packages(callback) {
+  this.models.PackageCache.findAll({ partitioner: 'cached' }, callback);
 };
 
-/**
- * Lookup Object for parsing build heads
- * @constructor
- * @param {Object} pkg - package object
- * @param {Object} heads - build heads for a given package in the fetched env
- */
-function Lookup(pkg, heads) {
-  this.pkg = pkg;
-  this.heads = heads;
-  this.latest = this.order(heads);
-}
-
-/**
- * Compute and order the heads to return the latest version
- * @function order
- * @param {Array} heads - Array of head objects
- * @returns {String} latest version of the set
- */
-Lookup.prototype.order = function order(heads) {
-  return heads.filter(Boolean)
-    .sort((a, b) => {
-      return semver.lt(a.version, b.version) ? -1 : 1;
-    }).reduce((latest, v) => {
-      return semver.lt(latest, v.version)
-        ? v.version
-        : latest;
-    }, '0.0.0');
-};
-
-/**
- * Compute objects to be sent over nsq for missing builds
- * @function missing
- * @returns {Array} of spec objects to be sent to nsq
- */
-Lookup.prototype.missing = function () {
-  return this.heads.map((head) => {
-    return semver.lt(head.version, this.latest)
-      ? this.specify(head)
-      : null;
-  }).filter(Boolean);
-};
-
-/**
- * Create a spec object given a head object and internal data
- * @function specify
- * @param {Object} head - A given head object from a missing build
- * @returns {Spec} object to be sent to nsq
- */
-Lookup.prototype.specify = function specify(head) {
-  return new Spec(
-    head.name,
-    head.env,
-    this.latest,
-    head.locale,
-    this.pkg.extended.build
-  );
-};
-
-/**
- * Spec Object
- * @constructor
- * @param {String} name - name of package
- * @param {String} env - env of package
- * @param {String} version - version of package
- * @param {String} locale - locale of package
- * @param {String} type - build type
- */
-function Spec(name, env, version, locale, type) {
-  this.name = name;
-  this.env = env;
-  this.version = version;
-  this.locale = locale;
-  this.type = type;
-}
 
 /**
  * scheduler preboot
  * @param {slay.App} app Slay application
  * @param {Object} options Additional configuration
- * @param {Function} next Continuation function
+ * @param {Function} callback Continuation function
  */
 module.exports = function schedboot(app, options, callback) {
   app.scheduler = new Scheduler(assign({
@@ -230,5 +160,3 @@ module.exports = function schedboot(app, options, callback) {
 };
 
 module.exports.Scheduler = Scheduler;
-module.exports.Spec = Spec;
-module.exports.Lookup = Lookup;
